syswit/requirements.txt
argparse
datetime
dcc
dash-core-components
plotly-express
py-libnuma
more-itertools
dash_bootstrap_components
dash==2.17.0
dash_bootstrap_components==1.6.0
jsonmerge==1.9.2
netifaces==0.11.0
pandas==2.2.2
plotly==5.22.0
psutil==5.9.8
PyYAML==6.0.1
setuptools==59.6.0
